Use schema information for all REST operations (#374)
Previously, the provider computed a resource ID during the Create operation, and then used this ID with hardcoded HTTP verbs for Read, Update, and Delete operations. While this works for many resources, some don’t follow this pattern. This change refactors the provider logic to use the endpoints and HTTP verbs specified for each operation by the schema.