Check for override functions on resource delete
Previously, the provider always mapped resource operations 1:1 with a REST API call. While this works for many cases, some operations are more complex, and require special handling. This change adds lookup tables of functions for each type of operation. If a function is present for a particular resource, this function will be executed by the provider instead of using the default REST logic.