Fix Conditions handling for IAM policy updates (#696)
The `getIamPolicy` endpoints are inconsistent across resources, and some include the `optionsRequestedPolicyVersion` query parameter. The version must be set to `3` to use IAM Conditions, so this parameter needs to be set on the endpoints that support it. Unfortunately, we can't include this parameter unconditionally because the request fails validation on the server for APIs that don't include it.