Remove validateOnly from SDKs and resource URLs (#866)
Previously, validateOnly was a property of the eventarc.Trigger resource (and some others). However, this property is an API implementation detail: API client may specify it to get a "dry run" without the actual action. See https://google.aip.dev/163#guidance

This PR ignores all query parameters with this specific name while building SDKs and resource URLs. As a result, we will never set it and it will stay false by default.

Fix #865